+++ CHANGES.md
@@ -1,1 +1,4 @@
 # Release Notes:
+- Docs improvement.
+- The tokens added to local state and prevented copmponent rerender with shouldComponentUpdate hook.
+- Added csrf support.